Moved collating and recursive .to to a new file
